requirements.txt
# ============================================================
# Dual-Agent Composite Hedge Trading System
# requirements.txt  —  Python 3.12+
# ============================================================

# ── Core async runtime ──────────────────────────────────────
asyncio-throttle==1.0.2
aiohttp==3.9.5
aiofiles==23.2.1
uvloop==0.19.0

# ── Exchange connectivity ────────────────────────────────────
ccxt==4.3.60
ccxt[async_support]

# ── Agent frameworks ─────────────────────────────────────────
langgraph==0.2.28
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-openai==0.1.23
crewai==0.63.0
crewai-tools==0.12.1

# ── LLM providers ───────────────────────────────────────────
openai==1.45.0
anthropic==0.34.2
# xai-sdk  # Grok — install from xAI when GA

# ── Data & numeric ───────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
ta==0.11.0                  # technical-analysis indicators
pandas-ta==0.3.14b0

# ── WebSocket streaming ───────────────────────────────────────
websockets==12.0
websocket-client==1.8.0

# ── Vector / memory DB ───────────────────────────────────────
chromadb==0.5.5
sentence-transformers==3.0.1

# ── Relational DB ─────────────────────────────────────────────
sqlalchemy==2.0.32
alembic==1.13.2
aiosqlite==0.20.0
asyncpg==0.29.0             # postgres async driver (optional)

# ── Config & env ──────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0

# ── Social / on-chain data sources ────────────────────────────
tweepy==4.14.0              # X/Twitter API v2
requests==2.32.3
httpx==0.27.2

# ── Logging & monitoring ──────────────────────────────────────
loguru==0.7.2
prometheus-client==0.20.0
opentelemetry-sdk==1.26.0

# ── Dashboard ─────────────────────────────────────────────────
streamlit==1.38.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
plotly==5.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4


# ── Defense Swarm ──────────────────────────────────────────────
# All defense agents use existing libs — no new runtime deps needed.
# Optional: install aiohttp-socks for SOCKS5 proxy rotation support
# aiohttp-socks==0.8.4
# Optional: install PySocks for socket-level proxy if IP blocking detected
# PySocks==1.7.1


# ── Testing & dev ─────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
black==24.8.0
ruff==0.6.4
mypy==1.11.2
